Namespace change and fixes
