+++ Content.Server/StationEvents/Components/VentCrittersRuleComponent.cs
@@ -1,4 +1,5 @@
-﻿using Content.Server.StationEvents.Events;
+﻿using Content.Server._Moffstation.StationEvents.Events;
+using Content.Server.StationEvents.Events;
 using Content.Shared.Storage;
 using Robust.Shared.Map;
 
@@ -16,14 +17,13 @@ public sealed partial class VentCrittersRuleComponent : Component
     [DataField("specialEntries")]
     public List<EntitySpawnEntry> SpecialEntries = new();
 
-
     // Moffstation - Start - New variables for single vent spawn
-    public EntityCoordinates? Location;
-
     /// <summary>
     /// The amount of chances something gets to spawn. estimated number of spawns can be calculated with (SpawnChances * entryProb)
     /// </summary>
     [DataField]
-    public int SpawnChances = 100;
+    public int SpawnAttempts = 100;
+
+    public EntityCoordinates? Location;
     // Moffstation - End
 }
